react 12장 useRef로 컴포넌트 안의 변수 만들기 정리 완료
