+++ React/01.리액트_입문/react-tutorial/begin-react/src/App.js
@@ -1,3 +1,4 @@
+import React, {useRef} from 'react';
 import Counter from './Counter';
 import Hello from './Hello'; // 뒤에 .js는 생략 가능
 import InputSample from './InputSample';
@@ -5,8 +6,33 @@ import UserList from './UserList';
 import Wrapper from './Wrapper';
 
 function App() {
+  const users = [
+    {
+        id: 1,
+        username: 'velopert',
+        email: 'public.velopert@gmail.com'
+    },
+    {
+        id: 2,
+        username: 'tester',
+        email: 'tester@example.com'
+    },
+    {
+        id: 3,
+        username: 'liz',
+        email: 'liz@example.com'
+    }
+  ];
+
+  const nextId = useRef(4);
+
+  const onCreate = () => {
+
+    console.log(nextId.current);
+  }
+
   return (
-    <UserList/>
+    <UserList users={users}/>
   );
 }
