+++ React/01.리액트_입문/react-tutorial/begin-react/src/UserList.js
@@ -8,25 +8,7 @@ function User({user}) {
     );
 }
 
-function UserList() {
-    const users = [
-        {
-            id: 1,
-            username: 'velopert',
-            email: 'public.velopert@gmail.com'
-        },
-        {
-            id: 2,
-            username: 'tester',
-            email: 'tester@example.com'
-        },
-        {
-            id: 3,
-            username: 'liz',
-            email: 'liz@example.com'
-        }
-    ];
-
+function UserList({users}) {
     return (
         <div>
             {
